# visual-compiler/backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.IDRequest:
    properties:
      project_name:
        description: User's project name
        type: string
      users_id:
        description: Represents the User's ID from frontend
        type: string
    required:
    - project_name
    - users_id
    type: object
...
  handlers.TranslatorRules:
    properties:
      project_name:
        description: User's project name
        type: string
      translation_rules:
        description: Translation Rules
        items:
          $ref: '#/definitions/services.TranslationRule'
        type: array
    required:
    - project_name
    - translation_rules
    type: object
...
  services.TranslationRule:
    properties:
      sequence:
        items:
          type: string
        type: array
      translation:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Documentation for Visual Compiler API endpoints
  title: Visual Compiler API
  version: "1.0"
paths:
...
  /translating/rules:
    post:
      consumes:
      - application/json
      description: Takes the user's rules, reads it, verifies it and stores it with
        the user's ID. If translation rules already exist, it overwrites those rules
        and removes any other created fields (translation)
      parameters:
      - description: Read Translation Rules From User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.TranslatorRules'
      produces:
      - application/json
      responses:
        "200":
          description: Translation Rules successfully stored
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input/Translation rules failed to insert
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create translation rules
      tags:
      - Translating
  /translating/translate:
    post:
      consumes:
      - application/json
      description: Searches the database for the user's Syntax tree and Translation
        Rules. If found, both are used to translate the tree into code. The code is
        either created or ,if already existing, updated. If the syntax tree and translation
        rules are not found, returns an error
      parameters:
      - description: Create Code from Syntax Tree and Translation Rules
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.IDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Code successfully created and stored
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input/Conversion failed
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Tree not found/Rules not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Translates syntax tree using the translation rules into code
      tags:
      - Translating
...
swagger: "2.0"